refactor(core): split parse.rs into document module (phase 1/5)
Mechanical split of crates/core/src/parse.rs (2,925 lines) into a
document module with named responsibilities. Zero behavior change:
all strings, error messages, error codes, and diagnostics are
byte-identical.

New layout:
  crates/core/src/document/
    mod.rs       — re-exports ParsedDocument, ParseOutput, BODY_FIELD
    fences.rs    — Lines, is_fence_marker_line, code_fence_on_line,
                   find_metadata_blocks, fence_opener_len
    sentinel.rs  — is_valid_tag_name, first_content_key, extract_sentinels
    assemble.rs  — MetadataBlock, build_block, decompose,
                   decompose_with_warnings, missing_quill_message
    limits.rs    — MAX_YAML_DEPTH (moved here from error.rs)
    tests/
      mod.rs
      assemble_tests.rs  — all test bodies from parse.rs, unchanged

Deleted: crates/core/src/parse.rs

error.rs now re-exports MAX_YAML_DEPTH from document::limits so
external callers using quillmark_core::error::MAX_YAML_DEPTH are
unaffected.

lib.rs replaces `pub mod parse;` with `pub mod document;` and
re-exports the same three public names.

Internal references in normalize.rs and quillmark/tests/security_tests.rs
updated from crate::parse:: / quillmark_core::parse:: to use the
public re-exports or crate::document::.
