+++ crates/core/src/document/tests/mod.rs
@@ -1,0 +1,1 @@
+mod assemble_tests;
